fix: move event sourcing union identifier tests to the package that owns them
CI runs each package against only its own declared dependencies, so the
event-sourcing tests failed there even though they passed locally, where
every package shares one vendor tree. packages/Ecotone has no event
sourcing dependency and packages/JmsConverter had none either, so both
the failing-without-discriminator and working-with-discriminator cases
now live in JmsConverter, which declares ecotone/pdo-event-sourcing as a
require-dev dependency for exactly this kind of integration test.
