+++ packages/JmsConverter/composer.json
@@ -54,6 +54,7 @@
         "symfony/cache": "^6.4|^7.0|^8.0"
     },
     "require-dev": {
+        "ecotone/pdo-event-sourcing": "~1.320.0",
         "phpunit/phpunit": "^10.5|^11.0",
         "phpstan/phpstan": "^1.8"
     },
