+++ packages/JmsConverter/tests/Fixture/ExamplesToConvert/UnionType/InternalId.php
@@ -15,4 +15,9 @@ final class InternalId
     {
         return $this->id;
     }
+
+    public function __toString(): string
+    {
+        return $this->id;
+    }
 }
