+++ packages/JmsConverter/tests/Fixture/UnionIdentifierAggregateWithoutDiscriminator/CreateTicket.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Test\Ecotone\JMSConverter\Fixture\UnionIdentifierAggregateWithoutDiscriminator;
+
+use Test\Ecotone\JMSConverter\Fixture\ExamplesToConvert\UnionType\ExternalId;
+use Test\Ecotone\JMSConverter\Fixture\ExamplesToConvert\UnionType\InternalId;
+
+/**
+ * licence Apache-2.0
+ */
+final class CreateTicket
+{
+    public function __construct(public readonly InternalId|ExternalId $ticketId)
+    {
+    }
+}
